Separate `main` and `run`
